refactor: split core.main responsibilities across submodules
refactor: core
